ocr-processor/requirements.txt
requests
Pillow
python-dotenv
pdf2image

